Add files via upload
The new files were too big to upload, so run the code. The values correspond to "MSS" : 0,"MSI-H" : 1,"MSI-L" : 2, nan : 3. (we have to handle the missing values). The code creates 3 files, one for the new features one for the target variables and one where  the feature and target variable data are combined into a single CSV file.